refactor: CodeBlock
